requirements.txt
python_dotenv==1.0.0
python-telegram-bot==13.*
redis==4.5.2